Refactor package entrypoints into barrels
