+++ .changeset/refactor-package-barrels.md
@@ -1,0 +1,8 @@
+---
+"@anvia/fastembed": patch
+"@anvia/transformers": patch
+"@anvia/langfuse": patch
+"@anvia/otel": patch
+---
+
+Refactor package entrypoints into barrel exports with focused internal modules.
